Got OSI emulation to work (kinda)
Also changed source files from UTF-16 to ANSI and removed diagrams that were drawn with Propeller characters
This version of the OSI emulator starts up and runs, but the keyboard still apears a little glitchy. The ACIA emulation is not implemented yet.
The reason I couldn't get the keyboard working before is that the original hardware used negative logic and our emulator used positive logic. I fixed it by putting two xor's in the keyboard access cog for now, but that made it go over the timing budget which may be what's causing the glitchiness.