Improving keyboard code
The code in the PS/2 keyboard driver has always bugged me a lot; it depended on "magic numbers" scattered all across the code. This is an attempt to fix it by replacing fields in the VAR section by an array with index constants.
The normal functionality of getting keys from the keyboard in Spin hasn't been tested, but at least now the OSI keyboard matrix is mostly stable, though it seems that sometimes the shift and caps lock status appears to get out of sync from the keyboard. I'll have to write a test in 6502 assembly and see what's going on. For now, this is enough to type short programs into the OSI emulator anyway.