Adding "Kimple1"
The Kimple1 is a mad mutation of the Apple 1 replica software and the MicroKim: it maps the Woz Mon / Apple Basic / Krusader ROM into the high address area, but enables the MicroKim on-board address decoder for addresses below the 8K border. That way, the RAM, RIOT and ROM are visible as usual on $0000-$1FFF and can be studied and controlled with Woz mon.