Adding MicroPle1 directory and updating readme
As third experiment in the MicroKim Lobotomy project, I modified the address decoder PASM code a little bit so it only enables the on-board address decoder for the RAM and the RIOT, but not for the EPROM (i.e. only for addresses $0000-$17FF). Also added an extra memory cog instance to map the KIM-1 ROMs in their normal location $1800-$1FFF.

The result is a mutated Apple-1 / Kim-1 combination, where you can use Woz Mon, Apple Basic, Krusader as usual, but you can also use them to download and program Kim programs. You can use the keys, the display, the KIM ROM routines and everything. But you can also use the Kim monitor to see the code in the Woz monitor and of course start it by typing AD F F 0 0 GO.

Note: single stepping through the Apple ROM is probably possible but you would have to change the interrupt vectors by patching them in the Propeller program. As an example, I already patched the reset vector so it jumps straight into the KIM ROM ($1C22).