Added a slider switch to the PCB, and reduced board size
The switch that I selected from Mouser didn't quite fit on the board with the components arranged the way they were, so I had to move a lot of stuff around. As a matter of fact, I had to move a LOT of stuff around and decided to try and make the board smaller. I would have liked to get the size under 100mm x 80mm (1/2 Eurocard) but didn't quite make it. Still, the new board is 13% smaller than the old size.