Adding initial version of Apple1 with SRAM contol
I implemented the SRAM control module (make sure you set the jumpers as indicated in the Hardware module) and changed the rest of the code to use the 1-pin keyboard from OBEX. Unfortunately the 1-pin keyboard needs a key press to initialize the timing, so I made it possible to disable the keyboard in the Serial/keyboard module by setting the keyboard data pin to a negative number. The current implementation disables the keyboard, you can only enter input via the serial port.
This has not been tested yet.