Fixing minor bugs to Apple1SRAM project. Still doesn't work.
At least now the Apple 1 emulator starts so I can use Woz Mon to help me debug this.
Apparently it's possible to read and write the first SRAM memory location (configured at $2000) but for all other locations ($2001-$7FFF), Woz mon shows the upper half of the address as contents, which is an indication that the RAM chip isn't being enabled for those addresses.