Wrapping up the Apple1SRAM project
The SRAM control cog now works, so I changed the main module of the project to use SRAM for the entire RAM area. I used 32K because that's what the Replica 1 has, but it can be set to any multiple of 16, up to 53248.
The SRAM interferes with the automatic fake-reset feature of the memory cog so I changed the code not to use that anymore. Sorry folks, it'll be necessary to push RESET6502 for this.