Fixed jumpers, added Reset, moved LED
I inadvertently put ~RAMEN on the wrong jumper in Rev2 (J2 instead of J4) so it was impossible to have keyboard, video and SRAM. This moves ~RAMEN to J4 and A16 to J2.
Also added ~RESET6502 to the jumpers to make it possible to connect a Propeller pin to it
There was not enough space around the RESET6502 button in Rev2 so I moved the LED to make space. I also moved both pushbutton switches to be parallel and centered between surrounding parts.
Also some more minor modifications to PCB and schematic.