Rev.4 schematic and PCB
Replaced the expansion connector by a 50-pin header so I could add all the signals including BE (I added a pull-up resistor for the BE pin). It's even possible to replicate the entire jumper block on an expension board or an external connector now. Another possibility is to cut off the entire right side of the board (with all the connections and the jumper block) or leave it unpopulated, and make the necessary connections via the expansion port. The new connector is backwards compatible with the old one from the Rev.1, Rev.2 and Rev.3 boards.

Rewired almost the entire board to make it smaller again; I finally got it to fit into slightly over 80x100mm (you'll have to cut a small sliver off the top or bottom to make it fit completely)

For this, I slightly reduced the size of the TO220 pads for the voltage regulators.

Also fixed a small problem in the mini-din 6 footprint: the shield is now grounded.