Nudged the 3.3V Electrolytic over a bit
When soldering the Rev.5 PCB, the only small problem was that the two electrolytic capacitors were just a little too close together. They still fit but space is tight, so I moved the one that's closest to the SRAM chip over a bit (off the grid). Nothing else was changed so this PCB is ready for production.