Modified Apple1 project for use with PropellerIDE
PropellerIDE (really BSTC and OpenSpin) apparently look for included Spin files (OBJ section) in the library before looking in the current directory, which caused a problem with "clock.spin" because a module with that name also exists in the library. I renamed the module to work around the problem.

Also, PropellerIDE generates files with .binary extensions which I added to .gitignore.

Also updated copyright year of the Apple1.spin module and added a patch to make the degree symbol go away. The degree symbol is shown when Woz mon tries to set the direction register of the PIA, but in our emulation, there is no such register and it doesn't need to be set. So the patch redirects the STY instruction from $D012 to $D013 which has no effect.