Changed OSI video driver to show all characters in video Memory
The video driver was once the start of the cooperation between Vince Briel and I. He didn't know how to modify the 1 pin video driver for his Superboard III project, and I didn't know how the original video hardware of the Superboard II worked. There were a few misunderstandings and long story short, the video driver showed a 25x25 character window that would be used by BASIC. But actually the original hardware shows the entire 32x32 matrix and relies on the software not to use the memory locations that are at the edge of the screen.
This commit fixes that: the entire matrix is shown, as much as possible anyway. In NTSC, there aren't enough video lines to show the bottom 4 text lines of the 32 row matrix.