Added PET-2001 project (doesn't work yet)
This is the "first sign of life" commit for the PET-2001 project.

The top level file (PET2001.spin) sets up the SRAM chip to represent 32KB RAM, and runs the 1-pin video module with a PET font and a memory module with the necessary ROMs for a PET-2001 with Basic-2 installed. There is no I/O yet: the area in memory where normally two PIAs and a VIA would be, is decoded to an 8K ROM wasteland. This doesn't make the Kernal happy: it drops out to the machine language monitor instead of showing "COMMODORE BASIC". There is no blinking cursor because there's no IRQ that should normally occur 60 or 50 times per second. And the MLM output is garbled; looks like there is bus contention between the emulated screen RAM and the SRAM or something. But what I see is definitely the monitor dumping registers.

Lots of stuff to do but it's a start!

There are many ROM image files, downloaded from the Zimmers FTP site, well-known in 6502 and Commodore land. I'll clean up the directory later.

Some modules were reused from elsewhere in the L-Star project, some other modules are new: I added a module called TraceClockGen.spin which can be used to replace the regular Clockgen.spin module to generate tracing information on the serial port of the Propeller. It uses the txx.spin module for this, which I wrote a while ago for the Propeller SPDIF decoder project. txx.spin is a fast serial port transmitter which can generate output in decimal, hex and binary, and much more. See http://obex.parallax.com/object/870 for more info about TXX.spin